+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add instructions for `PersistentNameID2Attribute` class
+- Add `RequiredAttributes` class
 
 ## [v1.0.0] - 2019-09-10