+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Improve `OpaqueSmartID` logging
+- Improve `OpaqueSmartID` error messages
 
 ## [v3.0.1] - 2020-07-10