+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fix bug in parseUserId()
+- Fix getters for SP NameID element
 
 ## [v3.0.2] - 2020-11-18